refactor: 第三步 - MessageHandler 引入 ConversationListener
- MessageHandler 添加 conversation_listener 字段
- ConversationChanged/TotalUnreadCountChanged/UserInputStatusChanged 同时通知 listener 和 EventBus
- publish_total_unread_count_changed() 同时通知 listener
- client.rs 创建 ConversationListener 并传给 MessageHandler

+++ rust/src/domain/listener/conversation.rs
@@ -6,7 +6,7 @@ pub struct ConversationListener {
     pub on_changed: ListenerSet<Vec<Conversation>>,
     pub on_deleted: ListenerSet<Vec<String>>,
     pub on_new: ListenerSet<Vec<Conversation>>,
-    pub on_total_unread_count_changed: ListenerSet<i32>,
+    pub on_total_unread_count_changed: ListenerSet<i64>,
     pub on_sync_started: ListenerSet<()>,
     pub on_sync_finished: ListenerSet<()>,
     pub on_sync_failed: ListenerSet<String>,
